fix: update fuel calculator components
