+++ src/frontend/components/FuelCalculator/widgets/FuelCalculatorEconomyPredict.tsx
@@ -45,8 +45,6 @@ export const FuelCalculatorEconomyPredict: React.FC<
   // Reverted to always 'live' mode as per user request
   const scenariosToShow = displayData?.targetScenarios || [];
 
-  // console.log('[EconomyPredict] Render', { mode, currentLap: fuelData?.currentLap, bufferLen: bufferedScenarios.length, showLen: scenariosToShow.length });
-
   if (!fuelData || !scenariosToShow || scenariosToShow.length === 0) {
     return (
       <div
